add permission validate curd

+++ src/murajaah/murajaah.service.ts
@@ -1,10 +1,11 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { ForbiddenException, Injectable, NotFoundException } from '@nestjs/common';
 import { PrismaService } from 'src/prisma_config/prisma.service';
 import { MurajaahDto } from './murajaah.dto';
 
 @Injectable()
 export class MurajaahServices {
-  constructor(private prisma: PrismaService) {}
+  constructor(private readonly prisma: PrismaService
+  ) {}
   //   post murajaah
   async createMurajaahInMoth(
     data: MurajaahDto,
@@ -23,7 +24,7 @@ export class MurajaahServices {
     return { message: 'Create Murajaah Success', createMurajaah };
   }
   //   edit murajaah
-  async editMurajaahInMonth(data: MurajaahDto, murajaahId: string) {
+  async editMurajaahInMonth(data: MurajaahDto, murajaahId: string, userReqId:number) {
     const murajaah = await this.prisma.murajaah.findFirst({
       where: {
         id: parseInt(murajaahId),
@@ -31,6 +32,9 @@ export class MurajaahServices {
     });
     if (!murajaah) {
       throw new NotFoundException('Murajaah Notfound');
+    }
+    if(userReqId.toString() !== murajaah.userId.toString()){
+      throw new ForbiddenException('ForbidenAcces Update This Resource');
     }
     const tilawahEdit = await this.prisma.murajaah.update({
       where: {
@@ -46,7 +50,7 @@ export class MurajaahServices {
 
   //   delete murajaah
 
-  async deleteTilawahInMonth(murajaahId: string) {
+  async deleteTilawahInMonth(murajaahId: string,userReqId:number) {
     const murajaah = await this.prisma.murajaah.findFirst({
       where: {
         id: parseInt(murajaahId),
@@ -55,6 +59,9 @@ export class MurajaahServices {
     if (!murajaah) {
       throw new NotFoundException('Murajaah NotFound');
     }
+     if (userReqId.toString() !== murajaah.userId.toString()) {
+       throw new ForbiddenException('ForbidenAcces Update This Resource');
+     }
     const deleteTilawah = await this.prisma.murajaah.delete({
       where: {
         id: parseInt(murajaahId),
